Refactored the mapper-client to its own file
