+++ apps/ids/mapper-client.h
@@ -1,0 +1,8 @@
+#ifndef __IDS_MAPPER_CLIENT_H__
+#define __IDS_MAPPER_CLIENT_H__
+
+#include "contiki.h"
+
+PROCESS_NAME(mapper_client);
+
+#endif
